+++ demo/PaginationDemo.js
@@ -12,7 +12,7 @@ export default class Demo extends React.Component {
 				<p>快速跳转到某一页:</p>
 				<Pagination showQuickJumper={true} onChange={onChange} total={500} />
 				<p>迷你版本:</p>
-				<Pagination className="mini" showQuickJumper={true} onChange={onChange} total={50} showSizeChanger={true} />
+				<Pagination className="mini" showQuickJumper={true} onChange={onChange} total={50} sizeOptions={[5, 10, 20]} pageSize={5} showSizeChanger={true} />
 				<p>简单地翻页:</p>
 				<Pagination simple onChange={onChange} total={50} />
 			</div>