+++ HISTORY.md
@@ -2,6 +2,12 @@
 
 ---
 
-## 0.1.3
+## 0.1.4
 
 `CHANGED` upgrade to react@0.14
+
+
+## 0.1.3
+
+`CHANGED` Forked from rc-pagination
+`CHNAGED` simple style changed